feat,refactor(user,lessee)添加用户租户修改功能,增加CURD判断条件
通过前端获取对象进行用户租户的信息修改，判断CURD时对象是否为空、在数据库中是否存在

+++ src/main/java/com/cucaqi/controller/LesseeController.java
@@ -3,6 +3,7 @@ package com.cucaqi.controller;
 import com.cucaqi.constants.HTTP;
 import com.cucaqi.entity.Lessee;
 import com.cucaqi.entity.Result;
+import com.cucaqi.entity.User;
 import com.cucaqi.service.ILesseeService;
 import jdk.vm.ci.meta.Constant;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,10 +31,10 @@ public class LesseeController {
      */
     @GetMapping
     public Result getLesseeList(){
-
         Result result=new Result();
         try {
             List<Lessee> lesseeList = iLesseeService.getLesseeList();
+            //获取数据库中所有租户
             result.setData(lesseeList);
             result.setCode(HTTP.SUCCESS);
             result.setMsg("查询成功！");
@@ -54,15 +55,22 @@ public class LesseeController {
     @PostMapping("/addLessee")
     public Result addLessee(@RequestBody Lessee lessee){
         Result result=new Result();
-        try {
-            boolean number= iLesseeService.save(lessee);
-            result.setCode(HTTP.SUCCESS);
-            result.setMsg("添加成功！");
+        if(lessee==null){
+            result.setCode(HTTP.SERVER_ERR);
+            result.setMsg("添加失败--添加的租户为空");
         }
-        catch (Exception e){
-            result.setCode(HTTP.NOT_FOUND);
-            result.setMsg("添加租户失败--"+e.getLocalizedMessage());
+        else {
+            try {
+                boolean number= iLesseeService.save(lessee);
+                result.setCode(HTTP.SUCCESS);
+                result.setMsg("添加成功！");
+            }
+            catch (Exception e){
+                result.setCode(HTTP.NOT_FOUND);
+                result.setMsg("添加租户失败--"+e.getLocalizedMessage());
+            }
         }
+
 
         return result;
     }
@@ -75,25 +83,52 @@ public class LesseeController {
     @DeleteMapping("/deleteLessee")
     public Result deleteLessee(@RequestBody Lessee lessee){
         Result result=new Result();
-        try{
-            int id=lessee.getId();
-            Integer res = iLesseeService.deleteLessee(id);
-            if(res==exist){
-                result.setMsg("删除失败--该租户不是无数据关联租户");
-                result.setCode(HTTP.SERVER_ERR);
-            }
-            else {
-                result.setCode(HTTP.SUCCESS);
-                result.setMsg("删除成功");
-            }
+        int id=lessee.getId();
+        if(iLesseeService.getById(id)==null){
+            result.setMsg("删除失败--该租户不存在");
+            result.setCode(HTTP.SERVER_ERR);
         }
-        catch (Exception e){
-            result.setCode(HTTP.NOT_FOUND);
-            result.setMsg("删除租户失败--"+e.getLocalizedMessage());
+        else{
+            try{
+                Integer res = iLesseeService.deleteLessee(id);
+                if(res==exist){
+                    result.setMsg("删除失败--该租户不是无数据关联租户");
+                    result.setCode(HTTP.SERVER_ERR);
+                }
+                else {
+                    result.setCode(HTTP.SUCCESS);
+                    result.setMsg("删除成功");
+                }
+            }
+            catch (Exception e){
+                result.setCode(HTTP.NOT_FOUND);
+                result.setMsg("删除租户失败--"+e.getLocalizedMessage());
 
+            }
         }
+
        return result;
     }
+    @PutMapping("/updateLessee")
+    public Result updateUser(@RequestBody Lessee lessee){
+        Result result=new Result();
+        int id=lessee.getId();
+        if(iLesseeService.getById(id)==null){
+            result.setMsg("修改失败--该租户不存在");
+            result.setCode(HTTP.SERVER_ERR);
+        }
+        else {
+            try {
+                iLesseeService.updateById(lessee);
+            }
+            catch (Exception e){
+                result.setCode(HTTP.NOT_FOUND);
+                result.setMsg("租户修改失败--"+e.getLocalizedMessage());
+            }
+        }
 
+        return  result;
+
+    }
 
 }
